+++ eNMS/models/services/data_validation/netmiko_commands.py
@@ -21,6 +21,7 @@ class NetmikoValidationService(ConnectionService):
     driver = db.Column(db.SmallString)
     use_device_driver = db.Column(Boolean, default=True)
     use_textfsm = db.Column(Boolean, default=False)
+    use_genie = db.Column(Boolean, default=False)
     fast_cli = db.Column(Boolean, default=False)
     timeout = db.Column(Float, default=10.0)
     delay_factor = db.Column(Float, default=1.0)
@@ -58,6 +59,7 @@ class NetmikoValidationService(ConnectionService):
                 netmiko_connection.send_command(
                     command,
                     use_textfsm=run.use_textfsm,
+                    use_genie=run.use_genie,
                     delay_factor=run.delay_factor,
                     expect_string=run.sub(run.expect_string, locals()) or None,
                     auto_find_prompt=run.auto_find_prompt,
@@ -66,10 +68,10 @@ class NetmikoValidationService(ConnectionService):
                 )
                 for command in commands.splitlines()
             ]
-            if run.use_textfsm:
-                result = result[0]
+            if len(result) == 1:
+                (result,) = result
             elif not run.results_as_list:
-                result = "\n".join(result)
+                result = "\n".join(map(str, result))
             run.exit_remote_device(netmiko_connection, prompt, device)
         except Exception:
             result = netmiko_connection.session_log.getvalue().decode().lstrip("\u0000")
@@ -87,6 +89,7 @@ class NetmikoValidationForm(NetmikoForm):
     commands = StringField(substitution=True, widget=TextArea(), render_kw={"rows": 5})
     results_as_list = BooleanField("Results As List")
     use_textfsm = BooleanField("Use TextFSM", default=False)
+    use_genie = BooleanField("Use Genie / PyATS", default=False)
     expect_string = StringField(substitution=True, help="netmiko/expect_string")
     config_mode_command = StringField(help="netmiko/config_mode_command")
     auto_find_prompt = BooleanField(default=True, help="netmiko/auto_find_prompt")
@@ -101,6 +104,7 @@ class NetmikoValidationForm(NetmikoForm):
         "Advanced Netmiko Parameters": {
             "commands": [
                 "use_textfsm",
+                "use_genie",
                 "expect_string",
                 "config_mode_command",
                 "auto_find_prompt",